% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_align.R
\name{align_modification}
\alias{align_modification}
\title{Align a modification to a peptide sequence}
\usage{
align_modification(peptide = NULL, modification = NULL)
}
\arguments{
\item{peptide}{a character string representing a peptide sequence}

\item{modification}{a character string representing a modification and location probability}
}
\value{
a tidyproteomics data-object
}
\description{
Align a modification to a peptide sequence
}
