% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_align.R
\name{align_peptide}
\alias{align_peptide}
\title{Align a peptide sequence to a protein sequence}
\usage{
align_peptide(peptide = NULL, protein = NULL)
}
\arguments{
\item{peptide}{a character string representing a peptide sequence}

\item{protein}{a character string representing a protein sequence}
}
\value{
a tidyproteomics data-object
}
\description{
Align a peptide sequence to a protein sequence
}
