% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_counts.R
\name{analysis_counts}
\alias{analysis_counts}
\title{A function for evaluating expression differences between two sample sets via the limma algorithm}
\usage{
analysis_counts(data = NULL, impute_max = 0.5)
}
\arguments{
\item{data}{tidyproteomics data object}

\item{impute_max}{a numeric representing the largest allowable imputation percentage}
}
\value{
a tibble
}
\description{
A function for evaluating expression differences between two sample sets via the limma algorithm
}
