% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_enrichments.R
\name{analyze_enrichments}
\alias{analyze_enrichments}
\title{Analysis tables and plots of expression values}
\usage{
analyze_enrichments(
  data = NULL,
  top_n = 50,
  significance_max = 0.05,
  enriched_up_color = "blue",
  enriched_down_color = "red",
  height = 6.5,
  width = 10
)
}
\arguments{
\item{data}{a character defining the column name of the log2 foldchange values.}

\item{top_n}{a numerical value defining the number of terms to display in the plot}

\item{significance_max}{a numeric defining the maximum statistical significance to highlight.}

\item{enriched_up_color}{a color to assign the up enriched values}

\item{enriched_down_color}{a color to assign the down enriched values}

\item{width}{a numeric}
}
\value{
a tidyproteomics data object
}
\description{
\code{analyze_enrichments()} is a GGplot2 implementation for plotting the expression differences
as foldchange ~ statistical significance. See also \code{plot_proportion()}.  This function can
take either a tidyproteomics data object or a table with the required headers.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)
hela_proteins \%>\%
   expression(knockdown/control) \%>\%
   analyze_expressions(log2fc_min = 0.5, significance_column = "p_value")

}
