% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_expressions.R
\name{analyze_expressions}
\alias{analyze_expressions}
\title{Analysis tables and plots of expression values}
\usage{
analyze_expressions(
  data = NULL,
  log2fc_min = 1,
  log2fc_column = "log2_foldchange",
  significance_max = 0.05,
  significance_column = "adj_p_value",
  labels_column = NULL,
  show_pannels = TRUE,
  show_lines = TRUE,
  show_fc_scale = TRUE,
  show_title = TRUE,
  show_pval_1 = TRUE,
  point_size = NULL,
  color_positive = "dodgerblue",
  color_negative = "firebrick1",
  height = 5,
  width = 8
)
}
\arguments{
\item{log2fc_min}{a numeric defining the minimum log2 foldchange to highlight.}

\item{log2fc_column}{a character defining the column name of the log2 foldchange values.}

\item{significance_max}{a numeric defining the maximum statistical significance to highlight.}

\item{significance_column}{a character defining the column name of the statistical significance values.}

\item{labels_column}{a character defining the column name of the column for labeling.}

\item{show_pannels}{a boolean for showing colored up/down expression panels.}

\item{show_lines}{a boolean for showing threshold lines.}

\item{show_fc_scale}{a boolean for showing the secondary foldchange scale.}

\item{show_title}{input FALSE, TRUE for an auto-generated title or any charcter string.}

\item{show_pval_1}{a boolean for showing expressions with pvalue == 1.}

\item{point_size}{a character reference to a numerical value in the expression table}

\item{color_positive}{a character defining the color for positive (up) expression.}

\item{color_negative}{a character defining the color for negative (down) expression.}

\item{height}{a numeric}

\item{width}{a numeric}
}
\value{
a tidyproteomics data object
}
\description{
\code{analyze_expressions()} is a GGplot2 implementation for plotting the expression differences
as foldchange ~ statistical significance. See also \code{plot_proportion()}.  This function can
take either a tidyproteomics data object or a table with the required headers.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)
hela_proteins \%>\%
   expression(knockdown/control) \%>\%
   analyze_expressions(log2fc_min = 0.5, significance_column = "p_value")

}
