% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_asdataframe.R
\name{as.data.frame.tidyproteomics}
\alias{as.data.frame.tidyproteomics}
\title{Helper function to convert the data-object into a tibble}
\usage{
\method{as.data.frame}{tidyproteomics}(data, shape = c("long", "wide"), values = NULL, drop = NULL)
}
\arguments{
\item{data}{tidyproteomics data object}

\item{shape}{the orientation of the quantitative data as either a single measure
per row (\strong{long}), or as multiple measures per protein/peptide (\strong{wide}).}

\item{values}{indicates the selected normalization to output. The default is
that selected at the time of normalization.}
}
\value{
a tibble
}
\description{
\code{as.data.frame()} is a function that converts the tidyproteomics data object into
a tibble. This tibble is in the long-format, such that a there is a single
observation per line.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)

# convert the data-object to a data.frame
hela_proteins \%>\% as.data.frame() \%>\% as_tibble()

# select the wide format
hela_proteins \%>\% as.data.frame(shape = 'wide') \%>\% as_tibble()

# select the wide format & drop some columns
hela_proteins \%>\%
   as.data.frame(shape = 'wide',
                 drop = c('description','wiki_pathway','reactome_pathway','biological_process')) \%>\%
   as_tibble()

}
