% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_span.R
\name{axis_breaks}
\alias{axis_breaks}
\title{Helper function for plotting a human readable axis}
\usage{
axis_breaks(x)
}
\arguments{
\item{data}{tidyproteomics data object}

\item{pairs}{the list of vector doublets}
}
\value{
list of vectors
}
\description{
Helper function for plotting a human readable axis
}
