% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_enrichment.R
\name{calc_enrichment}
\alias{calc_enrichment}
\title{Helper function to calculate term enrichment}
\usage{
calc_enrichment(data, x)
}
\arguments{
\item{data}{tidyproteomics data table object}

\item{x}{the annotation to compute enrichment for}
}
\value{
list of vectors
}
\description{
Helper function to calculate term enrichment
}
