% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_transform_center.R
\name{center}
\alias{center}
\title{helper function for normalizing a quantitative table}
\usage{
center(
  table,
  group_by = c("identifier"),
  values = "abundance",
  method = c("median", "mean", "geomean", "sum")
)
}
\arguments{
\item{table}{a tibble}

\item{group_by}{character vector}

\item{values}{character string}

\item{method}{character string}
}
\value{
a tibble
}
\description{
helper function for normalizing a quantitative table
}
