% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_check.R
\name{check_data}
\alias{check_data}
\title{Check the integrity of a tidyproteomics data object}
\usage{
check_data(data = NULL)
}
\arguments{
\item{data}{tidyproteomics data object}
}
\value{
silent on success, an abort message on fail
}
\description{
\code{check_data()} is a helper function that checks the structure and contents of
a tidyproteomics data object
}
