% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_pairs.R
\name{check_pairs}
\alias{check_pairs}
\title{Helper function for iterative expression analysis}
\usage{
check_pairs(pairs = NULL, sample_names = NULL)
}
\arguments{
\item{pairs}{the list of vector doublets}

\item{data}{tidyproteomics data object}
}
\value{
list of vectors
}
\description{
Helper function for iterative expression analysis
}
