% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_check.R
\name{check_table}
\alias{check_table}
\title{Check the integrity of a tidyproteomics quantitative tibble}
\usage{
check_table(table = NULL)
}
\arguments{
\item{table}{a tibble}
}
\value{
silent on success, an abort message on fail
}
\description{
\code{check_table()} is a helper function that checks the structure and contents of
a tidyproteomics quantitative tibble
}
