% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_codify.R
\name{codify}
\alias{codify}
\title{Build a tidyproteomics data object}
\usage{
codify(table = NULL, identifier = NULL, annotations = NULL)
}
\arguments{
\item{table}{tidyproteomics data object}

\item{identifier}{a character vector}

\item{annotations}{a character vector}
}
\value{
tidyproteomics data object
}
\description{
\code{data_codify()} is a helper function
}
