% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{data_import}
\alias{data_import}
\title{A helper function for importing peptide table data}
\usage{
data_import(file_names = NULL, platform = NULL, analyte = NULL, path = NULL)
}
\arguments{
\item{file_names}{a character vector of file paths}

\item{platform}{a character string}

\item{analyte}{a character string}

\item{path}{a character string}
}
\value{
a tidyproteomics list data-object
}
\description{
A helper function for importing peptide table data
}
