% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_subset.R
\name{down_select}
\alias{down_select}
\title{Helper function to subset a data frame}
\usage{
down_select(table = NULL, tidyproteomics_quo = NULL)
}
\arguments{
\item{table}{a tibble}

\item{tidyproteomics_quo}{a character vector}
}
\value{
a tibble
}
\description{
Helper function to subset a data frame
}
