% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoli_peptides}
\alias{ecoli_peptides}
\title{A sample tidyproteomics data object}
\format{
A list collection of character values and tibbles:
\describe{
\item{quantitative}{tibble, protein quantitative data}
\item{annotation}{tibble, protein annotation data}
...
}
}
\usage{
ecoli_peptides
}
\description{
A dataset containing the quantitative peptide data for ten proteins from 4
samples with 3 replicates each
}
\keyword{datasets}
