% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_enrichment.R
\name{enrichment}
\alias{enrichment}
\title{Compute protein enrichment}
\usage{
enrichment(
  data = NULL,
  ...,
  .pairs = NULL,
  .terms = NULL,
  .method = c("gsea", "wilcoxon", "fishers_exact"),
  .score_type = c("std", "pos", "neg"),
  .log2fc_min = 0,
  .significance_max = 0.05,
  .significance_min = .significance_max,
  .cpu_cores = 1
)
}
\arguments{
\item{data}{tidyproteomics data object}

\item{...}{two sample comparison e.g. experimental/control}

\item{.pairs}{a list of vectors each containing two named sample groups}

\item{.terms}{a character string referencing "term(s)" in the annotations table}

\item{.method}{a character string}

\item{.score_type}{a character string. From the fgsea manual: "This parameter
defines the GSEA score type. Possible options are ("std", "pos", "neg"). By
default ("std") the enrichment score is computed as in the original GSEA. The
"pos" and "neg" score types are intended to be used for one-tailed tests
(i.e. when one is interested only in positive ("pos") or negateive ("neg") enrichment)."}

\item{.significance_max}{used only for Fisher's Exact Test, a numeric defining the maximum statistical significance to consider as "enriched"}

\item{.significance_min}{deprecated alias for .significance_max}

\item{.cpu_cores}{the number of threads used to speed the calculation}
}
\value{
a tibble
}
\description{
\code{enrichment()} is an analysis function that computes the protein summary
statistics for a given tidyproteomics data object.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)

# using the default GSEA method
hela_proteins \%>\%
   expression(knockdown/control) \%>\%
   enrichment(knockdown/control, .terms = "biological_process") \%>\%
   export_analysis(knockdown/control, .analysis = "enrichment", .term = "biological_process")

# using a Wilcoxon Rank Sum method
hela_proteins \%>\%
   expression(knockdown/control) \%>\%
   enrichment(knockdown/control, .terms = "biological_process", .method = "wilcoxon") \%>\%
   export_analysis(knockdown/control, .analysis = "enrichment", .term = "biological_process")

# using the .pairs argument when multiple comparisons are needed
comps <- list(c("control","knockdown"),
            c("knockdown","control"))

hela_proteins \%>\%
   expression(.pairs = comps) \%>\%
   enrichment(.pairs = comps, .terms = c("biological_process", "molecular_function")

}
