% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_enrichment_fishers.R
\name{enrichment_fishersexact}
\alias{enrichment_fishersexact}
\title{A function for evaluating term enrichment via Fischer's Exact method}
\usage{
enrichment_fishersexact(
  data_expression = NULL,
  data = NULL,
  term_group = NULL,
  log2fc_min = 0,
  significance_min = 0.05,
  cpu_cores = 1,
  ...
)
}
\arguments{
\item{data_expression}{a tibble from and two sample expression difference analysis}

\item{data}{tidyproteomics data object}

\item{term_group}{a character string referencing "term" in the annotations table}

\item{log2fc_min}{a numeric defining the minimum log2 foldchange to consider as "enriched"}

\item{cpu_cores}{the number of threads used to speed the calculation}

\item{...}{pass through arguments}

\item{significance_max}{a numeric defining the maximum statistical significance to consider as "enriched"}
}
\value{
a tibble
}
\description{
A function for evaluating term enrichment via Fischer's Exact method
}
