% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_enrichment_wilcoxon.R
\name{enrichment_wilcoxon}
\alias{enrichment_wilcoxon}
\title{A function for evaluating term enrichment via Wilcoxon Rank Sum}
\usage{
enrichment_wilcoxon(
  data_expression = NULL,
  data = NULL,
  term_group = NULL,
  cpu_cores = 1,
  ...
)
}
\arguments{
\item{data_expression}{a tibble from and two sample expression difference analysis}

\item{data}{tidyproteomics data object}

\item{term_group}{a character string referencing "term" in the annotations table}

\item{cpu_cores}{the number of threads used to speed the calculation}

\item{...}{pass through arguments}
}
\value{
a tibble
}
\description{
A function for evaluating term enrichment via Wilcoxon Rank Sum
}
