% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_experimental.R
\name{experimental}
\alias{experimental}
\title{Returns the data experimental set up}
\usage{
experimental(data = NULL, destination = c("print", "save"))
}
\arguments{
\item{data}{tidyproteomics data object}

\item{destination}{a character string}
}
\value{
a character
}
\description{
\code{experimental()} returns the transformative operations performed on the data.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)
#\dontrun{
hela_proteins <- path_to_package_data("p97KD_HCT116") \%>\%
   import("ProteomeDiscoverer", "proteins") \%>\%
   reassign(sample == "ctl", .replace = "control") \%>\%
   reassign(sample == "p97", .replace = "knockdown") \%>\%
   impute() \%>\%
   normalize(.method = c("linear","loess"))
}
hela_proteins \%>\% experimental()

}
