% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_experimental.R
\name{experimental_groups}
\alias{experimental_groups}
\title{Main function for adding sample groups}
\usage{
experimental_groups(data = NULL, sample_groups = NULL)
}
\arguments{
\item{data}{a tidyproteomics data list-object}

\item{sample_groups}{a character string vector equal to the experimental row length}
}
\value{
a tidyproteomics data list-object
}
\description{
Main function for adding sample groups
}
