% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compexp.R
\name{export_compexp}
\alias{export_compexp}
\title{Comparative analysis between two expression tests}
\usage{
export_compexp(
  table_a = NULL,
  table_b = NULL,
  log2fc_min = 2,
  log2fc_column = "log2_foldchange",
  significance_max = 0.05,
  significance_column = "adj_p_value",
  labels_column = "protein",
  export = c("all", "a_only", "b_only", "intersect")
)
}
\arguments{
\item{table_a}{a tibble}

\item{table_b}{a tibble}

\item{log2fc_min}{a numeric defining the minimum log2 foldchange to highlight.}

\item{log2fc_column}{a character defining the column name of the log2 foldchange values.}

\item{significance_max}{a numeric defining the maximum statistical significance to highlight.}

\item{significance_column}{a character defining the column name of the statistical significance values.}

\item{labels_column}{a character defining the column name of the column for labeling.}

\item{export}{a character string for the significance data to return}
}
\value{
a tibble
}
\description{
\code{export_compexp()} returns a table of the comparison in
expression differences between two methods or two sets of groups. For example,
one could run an expression difference for two different conditions (A and B)
prodived the experiment contained 3 samples condition A, condition B and WT,
then compare those results. The proteins showing up in the intersection
indicate common targets for condition A and B.

\if{html}{\out{<div class="sourceCode">}}\preformatted{expdiff_a <- protein_data \%>\%
   expression(experiment = "condition_a", control = "wt")

expdiff_b <- protein_data \%>\%
   expression(experiment = "condition_b", control = "wt")

export_compexp(expdiff_a, expdiff_b, export = "intersect")

}\if{html}{\out{</div>}}
}
