% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_export.R
\name{export_quant}
\alias{export_quant}
\title{Export the quantitative data from an tidyproteomics data-object}
\usage{
export_quant(
  data = NULL,
  file_name = NULL,
  raw_data = TRUE,
  normalized = FALSE,
  scaled = c("none", "between", "proportion")
)
}
\arguments{
\item{data}{tidyproteomics data object}

\item{file_name}{character string vector}

\item{raw_data}{a boolean}

\item{normalized}{a boolean}

\item{scaled}{a boolean}
}
\value{
a tibble
}
\description{
\code{export_quant()} returns the main quantitative data object as a tibble with
\emph{identifier} as the designation for the measured observation.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)
ecoli_proteins \%>\%
   normalize(.method = "loess") \%>\%
   export_quant(file_name = "ecoli_quant_data.xlsx", normalized = "loess")

}
