% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_expression.R
\name{expression}
\alias{expression}
\title{Summarize the data}
\usage{
expression(data = NULL, ..., .method = stats::t.test)
}
\arguments{
\item{data}{tidyproteomics data object}

\item{...}{two sample comparison e.g. experimental/control}

\item{.method}{a two-distribution test function returning a p_value for the null
hypothesis. Example functions include t.test, wilcox.test, stats::ks.test,
additionally, the string \emph{"limma"} can be used to select from the limma
package to compute an empirical Bayesian estimation which performs better with
non-linear distributions and uneven replicate balance between samples.}
}
\value{
a tibble
}
\description{
\code{expression()} is an analysis function that computes the protein summary
statistics for a given tidyproteomics data object.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)

# simple t.test expression analysis
ecoli_proteins \%>\%
   expression(ko/wt) \%>\%
   export_analysis(ko/wt, .analysis = "expression")

# a wilcox.test expression analysis
ecoli_proteins \%>\%
   expression(ko/wt, .method = stats::wilcox.test) \%>\%
   export_analysis(ko/wt, .analysis = "expression")

# a one-tailed wilcox.test expression analysis
wilcoxon_less <- function(x, y) {
   stats::wilcox.test(x, y, alternative = 'less')
}
ecoli_proteins <- ecoli_proteins \%>\%
   expression(ko/wt, .method = stats::wilcox.test)

ecoli_proteins \%>\% export_analysis(ko/wt, .analysis = "expression")

# Note: the userdefined function is preserved in the operations tracking
ecoli_proteins \%>\% operations()

# limma expression analysis
ecoli_proteins \%>\%
   expression(ko/wt, .method = "limma") \%>\%
   export_analysis(ko/wt, .analysis = "expression")

}
