% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_expressiondiff_limma.R
\name{expression_limma}
\alias{expression_limma}
\title{Calculate expression differences between two-samples}
\usage{
expression_limma(data = NULL, experiment = NULL, control = NULL)
}
\arguments{
\item{data}{tidyproteomics data object}

\item{experiment}{a character string representing the experimental sample set}

\item{control}{a character string representing the control sample set}
}
\value{
a tibble
}
\description{
\code{expression_limma()} is a function for evaluating expression differences
between two sample sets via the limma algorithm
}
