% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_expressiondiff_test.R
\name{expression_test}
\alias{expression_test}
\title{A function for evaluating expression differences between two sample sets via the limma algorithm}
\usage{
expression_test(
  data = NULL,
  experiment = NULL,
  control = NULL,
  m.test = stats::t.test,
  ...
)
}
\arguments{
\item{data}{tidyproteomics data object}

\item{experiment}{a character string representing the experimental sample set}

\item{control}{a character string representing the control sample set}

\item{m.test}{a two-distribution test function returning a p_value for the null
hypothesis. Default is t.test. Example functions include t.test, wilcox.test,
stats::ks.test ...}

\item{...}{pass through arguments}
}
\value{
a tibble
}
\description{
A function for evaluating expression differences between two sample sets via the limma algorithm
}
