% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extract.R
\name{extract}
\alias{extract}
\title{Main function for extracting quantitative data from a tidyproteomics data-object}
\usage{
extract(data = NULL, values = NULL, na.rm = FALSE)
}
\arguments{
\item{data}{tidyproteomics data object}

\item{values}{character string vector}

\item{na.rm}{a boolean}
}
\value{
a tibble
}
\description{
Main function for extracting quantitative data from a tidyproteomics data-object
}
