% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_extract.R
\name{fasta_extract}
\alias{fasta_extract}
\title{Get the string defined by the regex}
\usage{
fasta_extract(string = NULL, regex = NULL)
}
\arguments{
\item{string}{a character}

\item{regex}{a list}
}
\value{
a list
}
\description{
\code{fasta_extract()} get the current string based on regex
}
