% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_regex.R
\name{fasta_regex}
\alias{fasta_regex}
\title{Get/Set the FASTA meta data regex}
\usage{
fasta_regex(params = NULL)
}
\arguments{
\item{params}{as list}
}
\value{
a list
}
\description{
\code{fasta_regex()} gets and sets the current regex patters to assist the \code{parse()} function.
This simply provides the structure needed to parse the fasta file, a custom list
can also be supplied. To set elements in the \code{regex()} function, simply provide a
list with complementary names to over-write the current list.
}
\examples{
#\dontrun{
fasta_regex(list("accession" = "sp\\\\|[A-Z]"))
}
}
