% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_aggregate.R
\name{fgeomean}
\alias{fgeomean}
\title{Calculates the geometric mean of a numeric vector with NAs removed}
\usage{
fgeomean(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a numeric
}
\description{
Calculates the geometric mean of a numeric vector with NAs removed
}
\examples{
library(tidyproteomics)
fgeomean(c(1,2,5,6,8,NA,NA))

}
