% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_aggregate.R
\name{fmean}
\alias{fmean}
\title{Calculates the mean of a numeric vector with NAs removed}
\usage{
fmean(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a numeric
}
\description{
Calculates the mean of a numeric vector with NAs removed
}
\examples{
library(tidyproteomics)
fmean(c(1,2,5,6,8,NA,NA))

}
