% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_aggregate.R
\name{fmin}
\alias{fmin}
\title{Calculates the minimum of a numeric vector with NAs removed}
\usage{
fmin(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a numeric
}
\description{
Calculates the minimum of a numeric vector with NAs removed
}
\examples{
library(tidyproteomics)
fmin(c(1,2,5,6,8,NA,NA))

}
