% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_accountings.R
\name{get_accountings}
\alias{get_accountings}
\title{Helper function to get all accounting terms}
\usage{
get_accountings(data = NULL)
}
\arguments{
\item{data}{tidyproteomics data object}

\item{term}{a character string}
}
\value{
a vector
}
\description{
Helper function to get all accounting terms
}
