% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annotations.R
\name{get_annotation_terms}
\alias{get_annotation_terms}
\title{Helper function to get available terms}
\usage{
get_annotation_terms(data)
}
\arguments{
\item{data}{tidyproteomics data object}
}
\value{
a vector
}
\description{
Helper function to get available terms
}
