% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annotations.R
\name{get_annotations}
\alias{get_annotations}
\title{Helper function to get all annotations for a given term}
\usage{
get_annotations(data = NULL, term = NULL)
}
\arguments{
\item{data}{tidyproteomics data object}

\item{term}{a character string}
}
\value{
a vector
}
\description{
Helper function to get all annotations for a given term
}
