% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quant_names.R
\name{get_quant_names}
\alias{get_quant_names}
\title{Get the quantitative value names}
\usage{
get_quant_names(data)
}
\arguments{
\item{data}{a tidyproteomics data-object}
}
\value{
a character vector
}
\description{
\code{get_quant_names()} is a helper function that returns the names for all of the
normalized quantitative values, such as \emph{raw}, \emph{linear}, \emph{loess}
}
\examples{
library(tidyproteomics)
get_quant_names(ecoli_proteins)

}
