% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_names.R
\name{get_sample_names}
\alias{get_sample_names}
\title{Helper function to get all sample names}
\usage{
get_sample_names(data = NULL)
}
\arguments{
\item{data}{tidyproteomics data object}
}
\value{
a vector
}
\description{
Helper function to get all sample names
}
