% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variable.R
\name{get_segment}
\alias{get_segment}
\title{Helper function to get available terms}
\usage{
get_segment(data = NULL, variable = NULL, .verbose = TRUE)
}
\arguments{
\item{data}{tidyproteomics data object}

\item{variable}{a character string}

\item{.verbose}{a boolean}
}
\value{
a character
}
\description{
Helper function to get available terms
}
