% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unique_variables.R
\name{get_unique_variables}
\alias{get_unique_variables}
\title{Helper function to get all sample names}
\usage{
get_unique_variables(data = NULL, variable = NULL)
}
\arguments{
\item{data}{tidyproteomics data object}

\item{variable}{a string character}
}
\value{
a vector
}
\description{
Helper function to get all sample names
}
