% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variable.R
\name{get_variables}
\alias{get_variables}
\title{Helper function to get available terms}
\usage{
get_variables(
  data = NULL,
  segment = c("experiments", "quantitative", "annotations", "accounting")
)
}
\arguments{
\item{data}{tidyproteomics data object}

\item{segment}{a character string}
}
\value{
a vector
}
\description{
Helper function to get available terms
}
