% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_subset.R
\name{\%like\%}
\alias{\%like\%}
\title{Helper function for subsetting}
\usage{
a \%like\% b
}
\arguments{
\item{a}{a dplyr tibble column reference}

\item{b}{a dplyr tibble column reference}
}
\value{
a character string
}
\description{
Helper function for subsetting
}
