% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_check.R
\name{hash_vector}
\alias{hash_vector}
\title{Create a crc32 hash on a vector}
\usage{
hash_vector(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a hash of x
}
\description{
\code{hash_vector()} is a helper function that returns a crc32 hash on a vector
}
