% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_munge.R
\name{import_charge}
\alias{import_charge}
\title{A helper function for importing peptide table data and dealing with charge groups}
\usage{
import_charge(tbl_data = NULL, tbl_config = NULL)
}
\arguments{
\item{tbl_data}{a table of imported data}

\item{tbl_config}{a table of config values}
}
\value{
a tibble
}
\description{
A helper function for importing peptide table data and dealing with charge groups
}
