% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_munge.R
\name{import_extract}
\alias{import_extract}
\title{A helper function for importing peptide table data}
\usage{
import_extract(tbl_data = NULL, tbl_config = NULL, remove = FALSE)
}
\arguments{
\item{tbl_data}{a table of imported data}

\item{tbl_config}{a table of config values}

\item{remove}{as boolean to determine if the extracted column name should change or copy to a new, retaining the old}
}
\value{
a tibble
}
\description{
A helper function for importing peptide table data
}
