% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_proteomediscoverer_peptides.R
\name{import_pd_peps}
\alias{import_pd_peps}
\title{A helper function for importing peptide table data from ProteomeDiscoverer}
\usage{
import_pd_peps(file_names = NULL)
}
\arguments{
\item{file_names}{a character vector of file paths}
}
\value{
a tidyproteomics list data-object
}
\description{
A helper function for importing peptide table data from ProteomeDiscoverer
}
