% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_proteomediscoverer_proteins.R
\name{import_pd_prots}
\alias{import_pd_prots}
\title{A helper function for importing protein table data from ProteomeDiscoverer}
\usage{
import_pd_prots(file_names = NULL)
}
\arguments{
\item{file_names}{a character vector of file paths}
}
\value{
a list data object
}
\description{
A helper function for importing protein table data from ProteomeDiscoverer
}
