% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_randomforest.R
\name{impute.randomforest}
\alias{impute.randomforest}
\title{Imputes missing values based on the missForest function}
\usage{
impute.randomforest(matrix = NULL, cores = 2)
}
\arguments{
\item{matrix}{a matrix with some NAs}

\item{cores}{the number of threads used to speed the calculation}
}
\value{
a matrix with imputed values
}
\description{
Imputes missing values based on the missForest function
}
