% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_randomforest.R
\name{impute_randomforest}
\alias{impute_randomforest}
\title{Imputes missing values based on the missForest function}
\usage{
impute_randomforest(table = NULL, minimum_to_impute = 0.25, cores = 2)
}
\arguments{
\item{table}{a tibble}

\item{minimum_to_impute}{the minimum ratio to impute at}

\item{cores}{the number of threads used to speed the calculation}
}
\value{
a tibble
}
\description{
Imputes missing values based on the missForest function
}
