% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_imputstats.R
\name{impute_ratio}
\alias{impute_ratio}
\title{Helper function for calculating imputation stats}
\usage{
impute_ratio(x)
}
\arguments{
\item{x}{a tibble}
}
\value{
list of vectors
}
\description{
Helper function for calculating imputation stats
}
