% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_venn.R
\name{intersect_venn}
\alias{intersect_venn}
\title{Helper function extracting a subset of proteins}
\usage{
intersect_venn(data = NULL, include = NULL, exclude = NULL)
}
\arguments{
\item{data}{the tidyproteomics data object}

\item{include}{the set of proteins contained within the intersection of these samples}

\item{exclude}{the set of proteins found in these samples to exclude}
}
\value{
a character string
}
\description{
Helper function extracting a subset of proteins
}
