% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_intersection.R
\name{intersection}
\alias{intersection}
\title{Create a data subset}
\usage{
intersection(data = NULL, .include = NULL, .exclude = NULL)
}
\arguments{
\item{data}{tidyproteomics data object}

\item{.include}{when exporting the "intersection" this is the set of proteins
contained within the intersection of these samples}

\item{.exclude}{when exporting the "intersection" this is the set of proteins
found in these samples to exclude}
}
\value{
a tibble
}
\description{
\code{intersection()} is a specalized function for sub-setting quantitative data
from a tidyproteomics data-object based data overlapping between sample groups.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)

# creates a subset of just the proteins found in 'control'
hela_proteins \%>\%
   subset(imputed == 0) \%>\%
   intersection(.include = c('control'), .exclude = c('knockdown'))

}
