% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_venn.R
\name{list_venn}
\alias{list_venn}
\title{Helper function Venn and Euler plots}
\usage{
list_venn(data = NULL, ...)
}
\arguments{
\item{data}{tidyproteomics data object}

\item{...}{pass through arguments}
}
\value{
list of vectors
}
\description{
Helper function Venn and Euler plots
}
