% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_data_io.R
\name{load_local}
\alias{load_local}
\title{Load project specific data}
\usage{
load_local(analyte = c("peptides", "proteins"))
}
\arguments{
\item{analyte}{a character string}
}
\value{
an tidyproteomics data object
}
\description{
\code{load_local()} is a simple function that loads the current project
tidyproteomics data object
}
\examples{
library(tidyproteomics)
# hela_proteins <- load_omics(analyte = "proteins")

}
