% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_reverselog_transform.R
\name{log_breaks}
\alias{log_breaks}
\title{Generate breaks for log transformation}
\usage{
log_breaks(n = 11)
}
\arguments{
\item{n}{a numeric for the approximate number of breaks}
}
\value{
a function that generates breaks
}
\description{
Generate breaks for log transformation
}
