% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_vectors.R
\name{match_vect}
\alias{match_vect}
\title{match a named vector to string vector}
\usage{
match_vect(un_vec, n_vec)
}
\arguments{
\item{un_vec}{an un-named vector}

\item{n_vec}{a named vector}
}
\value{
a named vector
}
\description{
match a named vector to string vector
}
