% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_codify.R
\name{meld}
\alias{meld}
\title{Meld a tidyproteomics data object into a single table}
\usage{
meld(data = NULL, single_quant_source = FALSE)
}
\arguments{
\item{data}{tidyproteomics data object}

\item{single_quant_source}{a boolean to indicate if only a single quantitative value should be reported}
}
\value{
a tibble
}
\description{
\code{data_meld()} is a helper function
}
