% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_merge.R
\name{merge_quantitative}
\alias{merge_quantitative}
\title{Helper function merging normalized data back into the main data-object}
\usage{
merge_quantitative(data = NULL, data_quant = NULL, values = "raw")
}
\arguments{
\item{data}{tidyproteomics data subset tibble}

\item{data_quant}{tidyproteomics data subset tibble}

\item{values}{character string vector}
}
\value{
a tibble
}
\description{
Helper function merging normalized data back into the main data-object
}
