% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extract.R
\name{munge_identifier}
\alias{munge_identifier}
\title{Main function for munging peptide data from an extracted tidyproteomics data-object}
\usage{
munge_identifier(
  data,
  munge = c("combine", "separate"),
  identifiers = c("protein", "peptide", "modifications")
)
}
\arguments{
\item{data}{tidyproteomics data object}

\item{munge}{character string vector (combine | separate)}

\item{identifiers}{a character vector of the identifiers}
}
\value{
a tibble
}
\description{
Main function for munging peptide data from an extracted tidyproteomics data-object
}
