% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_limma.R
\name{normalize_limma}
\alias{normalize_limma}
\title{Normalization function for a tidyproteomics data-object}
\usage{
normalize_limma(data = NULL)
}
\arguments{
\item{data}{tidyproteomics data object}
}
\value{
a tibble
}
\description{
Normalization function for a tidyproteomics data-object
}
