% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_scaled.R
\name{normalize_scaled}
\alias{normalize_scaled}
\title{Normalization function for a tidyproteomics data-object}
\usage{
normalize_scaled(data = NULL, data_centered = NULL)
}
\arguments{
\item{data}{tidyproteomics list data-object}

\item{data_centered}{a tibble of centered values used for normalization}
}
\value{
a tibble
}
\description{
Normalization function for a tidyproteomics data-object
}
