% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_operations.R
\name{operations}
\alias{operations}
\title{Returns the data transformations}
\usage{
operations(data = NULL, destination = c("print", "save"))
}
\arguments{
\item{data}{tidyproteomics data object}

\item{destination}{a character string}
}
\value{
a character
}
\description{
\code{operations()} returns the transformative operations performed on the data.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)
ecoli_proteins <- path_to_package_data("ecoli-hint_proteins") \%>\%
   import("ProteomeDiscoverer", "proteins") \%>\%
   reassign("sample", "hint_ko", "ko") \%>\%
   impute() \%>\%
   normalize(.method = c("linear","loess"))

ecoli_proteins \%>\% operations()

}
