% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_path.R
\name{path_to_package_data}
\alias{path_to_package_data}
\title{Helper function for displaying path to data}
\usage{
path_to_package_data(item = c("proteins", "peptides", "fasta"))
}
\arguments{
\item{item}{a character string}
}
\value{
print the table to console
}
\description{
Helper function for displaying path to data
}
