% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_object.R
\name{plot.tidyproteomics}
\alias{plot.tidyproteomics}
\title{Tidy-Quant data object plot definition}
\usage{
\method{plot}{tidyproteomics}(obj)
}
\arguments{
\item{obj}{tidyproteomics data object}
}
\value{
print object summary
}
\description{
Tidy-Quant data object plot definition
}
