% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compexp.R
\name{plot_compexp}
\alias{plot_compexp}
\title{Comparative analysis between two expression tests}
\usage{
plot_compexp(
  table_a = NULL,
  table_b = NULL,
  log2fc_min = 2,
  log2fc_column = "log2_foldchange",
  significance_max = 0.05,
  significance_column = "adj_p_value",
  labels_column = "protein",
  point_size = NULL,
  show_lines = TRUE,
  color_a = "dodgerblue",
  color_b = "firebrick1",
  color_u = "purple"
)
}
\arguments{
\item{table_a}{a tibble}

\item{table_b}{a tibble}

\item{log2fc_min}{a numeric defining the minimum log2 foldchange to highlight.}

\item{log2fc_column}{a character defining the column name of the log2 foldchange values.}

\item{significance_max}{a numeric defining the maximum statistical significance to highlight.}

\item{significance_column}{a character defining the column name of the statistical significance values.}

\item{labels_column}{a character defining the column name of the column for labeling.}

\item{point_size}{a numeric for changing the point size.}

\item{show_lines}{a boolean for showing threshold lines.}

\item{color_a}{a character defining the color for table_a expression.}

\item{color_b}{a character defining the color for table_b expression.}

\item{color_u}{a character defining the color for the union between both tables.}
}
\value{
a ggplot2 object
}
\description{
\code{plot_compexp()} is a GGplot2 implementation for plotting the comparison in
expression differences between two methods or two sets of groups. For example,
one could run an expression difference for two different conditions (A and B)
prodived the experiment contained 3 samples condition A, condition B and WT,
then compare those results. The proteins showing up in the union (purple)
indicate common targets for condition A and B.

\if{html}{\out{<div class="sourceCode">}}\preformatted{expdiff_a <- protein_data \%>\%
   expression(experiment = "condition_a", control = "wt")

expdiff_b <- protein_data \%>\%
   expression(experiment = "condition_b", control = "wt")

plot_compexp(expdiff_a, expdiff_b)

}\if{html}{\out{</div>}}
}
\examples{
library(ggplot2, warn.conflicts = FALSE)
library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)

# comparing two analytical methods, in substitute for two conditions
exp_a <- ecoli_proteins \%>\% expression(ko/wt) \%>\%
     export_analysis(ko/wt, .analysis = 'expression')

exp_b <- ecoli_proteins \%>\% expression(ko/wt, .method = "limma") \%>\%
     export_analysis(ko/wt, .analysis = 'expression')

plot_compexp(exp_a, exp_b, log2fc_min = 1, significance_column = 'p_value') +
     ggplot2::labs(x = "(log2 FC) Wilcoxon Rank Sum",
                   y = "(log2 FC) Emperical Bayes (limma)",
                   title = "Ecoli Hint KO ~ WT")

}
