% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_counts.R
\name{plot_counts}
\alias{plot_counts}
\title{Plot the variation in normalized values}
\usage{
plot_counts(
  data = NULL,
  accounting = c("protein_groups", "proteins", "peptides", "peptides_unique"),
  impute_max = 0.5,
  palette = "YlGnBu",
  ...
)
}
\arguments{
\item{data}{tidyproteomics data object}

\item{accounting}{character string}

\item{impute_max}{a numeric representing the largest allowable imputation percentage}

\item{palette}{a string representing the palette for scale_fill_brewer()}

\item{...}{passthrough for ggsave see \code{plotting}}
}
\value{
a (tidyproteomics data-object | ggplot-object)
}
\description{
\code{plot_counts()} is a GGplot2 implementation for plotting counting statistics.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)
hela_proteins \%>\% plot_counts()

}
