% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_enrichment.R
\name{plot_enrichment}
\alias{plot_enrichment}
\title{Bubble plot of enrichment values}
\usage{
plot_enrichment(
  data = NULL,
  ...,
  .term = NULL,
  enrichment_min = 1,
  enrichment_column = "enrichment",
  significance_max = 0.01,
  significance_column = "p_value",
  term_column = "annotation",
  size_column = "size",
  destination = "plot",
  height = 5,
  width = 8
)
}
\arguments{
\item{...}{two sample comparison}

\item{enrichment_min}{a numeric defining the minimum log2 enrichment to highlight.}

\item{enrichment_column}{a character defining the column name of enrichment values.}

\item{significance_max}{a numeric defining the maximum statistical significance to highlight.}

\item{significance_column}{a character defining the column name of the statistical significance values.}

\item{term_column}{a character defining the column name for labeling.}

\item{size_column}{a character defining the column name of term size.}

\item{destination}{a character string}

\item{height}{a numeric}

\item{width}{a numeric}

\item{table}{a tibble}
}
\value{
a ggplot2 object
}
\description{
\code{plot_enrichment()} is a GGplot2 implementation for plotting the enrichment values. This function can
take either a tidyproteomics data object or a table with the required headers.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(ggplot2, warn.conflicts = FALSE)
library(tidyproteomics)
hela_proteins \%>\%
   expression(knockdown/control, .method = stats::t.test) \%>\%
   enrichment(knockdown/control, .term = 'biological_process', .method = "wilcoxon") \%>\%
   plot_enrichment(knockdown/control, .term = "biological_process") +
   labs(title = "Hela: Term Enrichment", subtitle = "Knockdown ~ Control")


}
