% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_venn-euler.R
\name{plot_euler}
\alias{plot_euler}
\title{GGplot2 extension to plot a Euler diagram}
\usage{
plot_euler(data, ...)
}
\arguments{
\item{data}{a tidyproteomics data object}

\item{...}{passthrough for ggsave see \code{plotting}}
}
\value{
a (tidyproteomics data-object | ggplot-object)
}
\description{
GGplot2 extension to plot a Euler diagram
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)
hela_proteins \%>\%
   subset(match_between_runs == FALSE) \%>\%
   plot_euler()

hela_proteins \%>\%
   subset(match_between_runs == FALSE) \%>\%
   subset(cellular_component \%like\% "cytosol") \%>\%
   plot_euler()

}
