% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot a heatmap of quantitative values by sample}
\usage{
plot_heatmap(data = NULL, tag = NULL, row_names = FALSE, ...)
}
\arguments{
\item{data}{tidyproteomics data object}

\item{tag}{a character string}

\item{row_names}{a boolean}

\item{...}{passthrough for ggsave see \code{plotting}}
}
\value{
a (tidyproteomics data-object | ggplot-object)
}
\description{
\code{plot_heatmap()} is a pheatmap implementation for plotting the commonly
visualized quantitative heatmap according to sample. Both the samples and the
quantitative values are clustered and visualized.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)
hela_proteins \%>\%
  normalize(.method = c("scaled", "median", "linear", "limma", "loess")) \%>\%
  select_normalization() \%>\%
  plot_heatmap()

}
