% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_normalization.R
\name{plot_normalization}
\alias{plot_normalization}
\title{Plot normalized values}
\usage{
plot_normalization(data = NULL, ...)
}
\arguments{
\item{data}{tidyproteomics data object}

\item{...}{passthrough for ggsave see \code{plotting}}
}
\value{
a (tidyproteomics data-object | ggplot-object)
}
\description{
\code{plot_normalization()} is a GGplot2 implementation for plotting the normalization
effects visualized as a box plot.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)
hela_proteins \%>\%
  normalize(.method = c("scaled", "median", "linear", "limma", "loess")) \%>\%
  plot_normalization()

}
