% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_pca}
\alias{plot_pca}
\title{Plot PCA values}
\usage{
plot_pca(
  data = NULL,
  variables = c("PC1", "PC2"),
  labels = TRUE,
  label_size = 3,
  ...
)
}
\arguments{
\item{data}{tidyproteomics data object}

\item{variables}{a character vector of the 2 PCs to plot. Acceptable values
include (PC1, PC2, PC3 ... PC9). Default c('PC1','PC2').}

\item{labels}{a boolean}

\item{label_size}{a numeric}

\item{...}{passthrough for ggsave see \code{plotting}}
}
\value{
a (tidyproteomics data-object | ggplot-object)
}
\description{
\code{plot_pca()} is a GGplot2 implementation for plotting two principal components
from a PCA analysis, visualized as a scatter.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)
hela_proteins <- hela_proteins \%>\%
  normalize(.method = c("scaled", "median", "linear", "limma", "loess")) \%>\%
  select_normalization()

hela_proteins \%>\% plot_pca()

# a different PC set
hela_proteins \%>\% plot_pca(variables = c("PC2", "PC3"))

# a PC scree plot
hela_proteins \%>\% plot_pca("scree")

}
