% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_plot.R
\name{plot_protein}
\alias{plot_protein}
\title{Visualize mapped sequence data}
\usage{
plot_protein(
  mapped_data = NULL,
  protein = NULL,
  row_length = 50,
  samples = NULL,
  modifications = NULL,
  ncol = NULL,
  nrow = NULL,
  color_sequence = "grey60",
  color_modifications = c("red", "blue", "orange", "skyblue", "purple", "yellow"),
  show_modification_precent = TRUE
)
}
\arguments{
\item{mapped_data}{a tidyproteomics data-object, specifically of sequencing origin}

\item{protein}{a character string}

\item{row_length}{a numeric}

\item{samples}{a character string}

\item{modifications}{a character string}

\item{ncol}{a numeric}

\item{nrow}{a numeric}

\item{color_sequence}{a character string}

\item{color_modifications}{a character vector}

\item{show_modification_precent}{a boolean}
}
\value{
a list of protein mappings
}
\description{
Visualize mapped sequence data
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)

ecoli_protein_map <- ecoli_peptides \%>\%
   protein_map(fasta = path_to_package_data('fasta'))

ecoli_protein_map \%>\% plot_protein('P0A6Y8')

}
