% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quantrank.R
\name{plot_quantrank}
\alias{plot_quantrank}
\title{Plot the variation in normalized values}
\usage{
plot_quantrank(
  data = NULL,
  accounting = "protein",
  type = c("points", "lines"),
  show_error = TRUE,
  show_rank_scale = FALSE,
  limit_rank = NULL,
  display_filter = c("none", "log2_foldchange", "p_value", "adj_p_value"),
  display_cutoff = 1,
  palette = "YlGnBu",
  impute_max = 0.5,
  ...
)
}
\arguments{
\item{data}{tidyproteomics data object}

\item{accounting}{character string}

\item{type}{character string}

\item{show_error}{a boolean}

\item{show_rank_scale}{a boolean}

\item{limit_rank}{a numerical vector of 2}

\item{display_filter}{a numeric between 0 and 1}

\item{display_cutoff}{a numeric between 0 and 1}

\item{palette}{a string representing the palette for scale_fill_brewer()}

\item{impute_max}{a numeric representing the largest allowable imputation percentage}

\item{...}{passthrough for ggsave see \code{plotting}}
}
\value{
a (tidyproteomics data-object | ggplot-object)
}
\description{
\code{plot_quantrank()} is a GGplot2 implementation for plotting the variability in
normalized values, generating two facets. The left facet is a plot of CVs for
each normalization method. The right facet is a plot of the 95\%CI in abundance,
essentially the conservative dynamic range. The goal is to select a normalization
method that minimizes CVs while also retaining the dynamic range.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)
ecoli_proteins \%>\% plot_quantrank()

ecoli_proteins \%>\% plot_quantrank(type = 'lines')

ecoli_proteins \%>\% plot_quantrank(display_filter = 'log2_foldchange', display_cutoff = 1)

ecoli_proteins \%>\% plot_quantrank(limit_rank = c(1,50), show_rank_scale = TRUE)

}
