% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_save.R
\name{plot_save}
\alias{plot_save}
\title{Helper function for saving plots}
\usage{
plot_save(
  plot,
  data,
  file_name,
  destination = c("plot", "save", "png", "svg", "tiff", "jpeg"),
  height = 5,
  width = 8,
  ...
)
}
\arguments{
\item{plot}{a ggplot2 object}

\item{data}{a tidyproteomics data object}

\item{file_name}{a character string}

\item{destination}{a character string}

\item{height}{a numeric}

\item{width}{a numeric}

\item{...}{passthrough ggplot2::ggsave arguments}
}
\value{
a ggplot2 object
}
\description{
\code{plot_save} helper function
}
