% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variation_cv.R
\name{plot_variation_cv}
\alias{plot_variation_cv}
\title{Plot the variation in normalized values}
\usage{
plot_variation_cv(data = NULL, ...)
}
\arguments{
\item{data}{tidyproteomics data object}

\item{...}{passthrough for ggsave see \code{plotting}}
}
\value{
a (tidyproteomics data-object | ggplot-object)
}
\description{
\code{plot_variation_cv()} is a GGplot2 implementation for plotting the variability in
normalized values, generating two facets. The left facet is a plot of CVs for
each normalization method. The right facet is a plot of the 95\%CI in abundance,
essentially the conservative dynamic range. The goal is to select a normalization
method that minimizes CVs while also retaining the dynamic range.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)
ecoli_proteins \%>\%
  normalize(.method = c("scaled", "median", "linear", "limma", "loess")) \%>\%
  plot_variation_cv()

}
