% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variation_pca.R
\name{plot_variation_pca}
\alias{plot_variation_pca}
\title{Plot the PCA variation in normalized values}
\usage{
plot_variation_pca(data = NULL, ...)
}
\arguments{
\item{data}{tidyproteomics data object}

\item{...}{passthrough for ggsave see \code{plotting}}
}
\value{
a (tidyproteomics data-object | ggplot-object)
}
\description{
\code{plot_variation_pca()} is a GGplot2 implementation for plotting the variability in
normalized values by PCA analysis, generating two facets. The left facet is a plot of CVs for
each normalization method. The right facet is a plot of the 95\%CI in abundance,
essentially the conservative dynamic range. The goal is to select a normalization
method that minimizes CVs while also retaining the dynamic range.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)
hela_proteins \%>\%
  normalize(.method = c("linear", "loess", "randomforest")) \%>\%
  plot_variation_pca()

}
