% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_venn-euler.R
\name{plot_venn}
\alias{plot_venn}
\title{GGplot2 extension to plot a Venn diagram}
\usage{
plot_venn(data, ...)
}
\arguments{
\item{data}{a tidyproteomics data object}

\item{...}{passthrough for ggsave see \code{plotting}}
}
\value{
a (tidyproteomics data-object | ggplot-object)
}
\description{
GGplot2 extension to plot a Venn diagram
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)
ecoli_proteins \%>\%
   subset(match_between_runs == FALSE) \%>\%
   plot_venn()

}
