% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_object.R
\name{println}
\alias{println}
\title{Helper function for printing messages}
\usage{
println(name = "", message = "", pad_length = 15)
}
\arguments{
\item{name}{string}

\item{message}{string}

\item{pad_length}{string}
}
\value{
console print line
}
\description{
Helper function for printing messages
}
