% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_map_munge.R
\name{protein_map_munge}
\alias{protein_map_munge}
\title{Align a peptide data to protein sequences for visualization}
\usage{
protein_map_munge(
  mapped_data = NULL,
  protein = NULL,
  row_length = 50,
  samples = NULL,
  modifications = NULL
)
}
\arguments{
\item{mapped_data}{a tidyproteomics data-object, specifically of peptide origin}

\item{protein}{a character string}

\item{row_length}{a numeric}

\item{samples}{a character string}

\item{modifications}{a character string}
}
\value{
a plot munged list of protein mappings
}
\description{
Align a peptide data to protein sequences for visualization
}
