% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_data_io_mzTab.R
\name{read_mzTab}
\alias{read_mzTab}
\title{A helper function for importing peptide table data}
\usage{
read_mzTab(path = NULL, analyte = c("peptides", "proteins"))
}
\arguments{
\item{path}{a character string}

\item{analyte}{a character string}
}
\value{
a tidyproteomics list data-object
}
\description{
A helper function for importing peptide table data
}
