% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reassign.R
\name{reassign}
\alias{reassign}
\title{reassign the sample info}
\usage{
reassign(data = NULL, field = NULL, pattern = NULL, replace = NULL)
}
\arguments{
\item{data}{a tidyproteomics data-object}

\item{field}{a character string}

\item{pattern}{a character string}

\item{replace}{a character string}
}
\value{
a tidyproteomics data-object
}
\description{
\code{reassign()} enables editing of the sample descriptive in the experimental table.
This function will only replace the sample string and update the replicate number.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)

# check the experiment table
ecoli_proteins \%>\% summary("experiment")

# make the modification
ecoli_proteins \%>\%
   reassign(field = "sample", pattern = "ko", replace = "hint_ko") \%>\%
   summary("sample")

# reassign specific file_ids
ecoli_proteins \%>\%
   reassign(field = "sample_file", pattern = "f1", replace = "new") \%>\%
   reassign(field = "sample_file", pattern = "f2", replace = "new") \%>\%
   summary("sample")

}
