% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_rmmbr.R
\name{rm.mbr}
\alias{rm.mbr}
\title{Remove MBR from the dataset across segments}
\usage{
rm.mbr(data = NULL, ..., .groups = c("all", "sample"))
}
\arguments{
\item{data}{tidyproteomics data object}

\item{...}{a three part expression (eg. x == a)}

\item{.groups}{a character string}
}
\value{
a tibble
}
\description{
\code{rm.mbr()} function is designed to remove match_between_runs between segments.
This function will return a smaller tidyproteomics data-object.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)

hela_proteins \%>\%
   summary('sample')

hela_proteins \%>\%
   rm.mbr(.groups = 'sample') \%>\%
   summary('sample')

}
