% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_export.R
\name{save_local}
\alias{save_local}
\title{Store data locally}
\usage{
save_local(data = NULL)
}
\arguments{
\item{data}{tidyproteomics data object}
}
\value{
tidyproteomics data object
}
\description{
\code{save_local()} will save the tidyproteomics data-object in the local project,
based on the given type in the directory ./data/ as either proteins.rds or
peptides.rds. This is a \emph{passthrough} function as it returns the original
tidyproteomics data-object.
}
