% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_export.R
\name{save_table}
\alias{save_table}
\title{Write table data locally}
\usage{
save_table(table, file_name = NULL)
}
\arguments{
\item{table}{a tibble}

\item{file_name}{a file name with extensions one of (.csv, .tsv, .rds, .xlsx)}
}
\value{
a tibble
}
\description{
\code{save_table()} will save a summary tibble in the root directory of the
local project, based on the extension given in the file name. This is a
\emph{passthrough} function as it returns the original tibble.
}
\examples{
#\dontrun{
library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)
hela_proteins \%>\%
   expression(knockdown/control) \%>\%
   export_analysis(knockdown/control, .analysis = "expression") \%>\%
   save_table("expression_limma_ko_over_wt.csv")
}

}
