% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_normalization.R
\name{select_normalization}
\alias{select_normalization}
\title{Select a normalization method}
\usage{
select_normalization(data = NULL, normalization = NULL)
}
\arguments{
\item{data}{tidyproteomics data object}

\item{normalization}{a character string}
}
\value{
a tidyproteomics data-object
}
\description{
\code{select_normalization()} selects the best normalization method base on low
CVs, low PCA (PC1), and wide Dynamic Range. This is a \emph{passthrough} function
as it returns the original tidyproteomics data-object.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)
hela_proteins <- hela_proteins \%>\%
  normalize(.method = c("scaled", "median", "linear", "limma", "loess","randomforest")) \%>\%
  select_normalization()

}
