% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_vectors.R
\name{set_vect}
\alias{set_vect}
\title{set a named vector}
\usage{
set_vect(config = NULL, category = NULL)
}
\arguments{
\item{config}{a data.frame of configuration values}

\item{category}{a character string}
}
\value{
a named vector
}
\description{
set a named vector
}
