% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_annotate.R
\name{show_annotations}
\alias{show_annotations}
\title{Display the current annotation data}
\usage{
show_annotations(data, term = NULL)
}
\arguments{
\item{data}{tidyproteomics data object}

\item{term}{a character string}
}
\value{
a vector
}
\description{
Display the current annotation data
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)

ecoli_proteins \%>\% show_annotations()

ecoli_proteins \%>\% show_annotations('reactome_pathway')

}
