% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_contamination.R
\name{stats_contamination}
\alias{stats_contamination}
\title{Assess the relative amount of protein contamination}
\usage{
stats_contamination(data = NULL, pattern = "CRAP")
}
\arguments{
\item{data}{tidyproteomics data object}

\item{pattern}{character string, regular expression}
}
\value{
a tibble
}
\description{
\code{stats_contamination()} is an analysis function that can take a regular
expression as a means to assign subsets of proteins as contaminant.
}
