% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_print.R
\name{stats_print}
\alias{stats_print}
\title{Helper function for displaying data}
\usage{
stats_print(table, title = NULL)
}
\arguments{
\item{table}{a tibble}

\item{title}{a character string}
}
\value{
print the table to console
}
\description{
Helper function for displaying data
}
