% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_summary.R
\name{stats_summary}
\alias{stats_summary}
\title{Summarize the protein accounting}
\usage{
stats_summary(
  data,
  group_by = c("global", "sample", "replicate", "experiment")
)
}
\arguments{
\item{data}{tidyproteomics data object}

\item{group_by}{what to summarize}
}
\value{
a tibble
}
\description{
\code{stats_summary()} is an analysis function that computes the protein summary
statistics for a given tidyproteomics data object.
}
