% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_str_normalize.R
\name{str_normalize}
\alias{str_normalize}
\title{Normalize the column names in a tibble}
\usage{
str_normalize(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a vector
}
\description{
Normalize the column names in a tibble
}
