% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_subset.R
\name{subset}
\alias{subset}
\title{Create a data subset}
\usage{
subset(data = NULL, ..., rm.mbr = TRUE, .verbose = TRUE)
}
\arguments{
\item{data}{tidyproteomics data object}

\item{...}{a three part expression (eg. x == a)}

\item{rm.mbr}{a boolean}

\item{.verbose}{a boolean}
}
\value{
a tibble
}
\description{
\code{subset()} is the main function for sub-setting quantitative data from a tidyproteomics
data-object based on a regular expression and targeted annotation. This function
will return a smaller tidyproteomics data-object.

Note: \code{rm.mbr()} is run as default, this is to remove MBR proteins that may no
longer have the original "anchor" observation present.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)

# creates a subset of just Ribosomes, based on the string in the annotation
# protein_description
hela_proteins \%>\%
   subset(description \%like\% "Ribosome") \%>\%
   summary()

# creates a subset without Ribosomes
hela_proteins \%>\%
   subset(!description \%like\% "Ribosome") \%>\%
   summary()

}
