% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_svm.R
\name{svm_parallel}
\alias{svm_parallel}
\title{parallel compute function for randomforest}
\usage{
svm_parallel(df)
}
\arguments{
\item{df}{a tibble of raw and centered values}
}
\value{
a tibble
}
\description{
parallel compute function for randomforest
}
