% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_theme_palette.R
\name{theme_palette}
\alias{theme_palette}
\title{helper function for having nice colors}
\usage{
theme_palette(n = 16)
}
\value{
character vector of curated html colors
}
\description{
helper function for having nice colors
}
