% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_import.R
\name{tidyproteomics}
\alias{tidyproteomics}
\title{Tidy-Quant data object print definition}
\usage{
tidyproteomics(obj)
}
\arguments{
\item{obj}{tidyproteomics data object}
}
\value{
print object summary
}
\description{
Tidy-Quant data object print definition
}
