% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_stats.R
\name{tidyproteomics_summary}
\alias{tidyproteomics_summary}
\title{Helper function to summarize the data}
\usage{
tidyproteomics_summary(
  data,
  by = c("global"),
  destination = c("print", "save", "return"),
  limit = 25,
  contamination = NULL
)
}
\arguments{
\item{data}{tidyproteomics data object}

\item{by}{what to summarize}

\item{destination}{character string, one of (save, print)}

\item{limit}{a numeric to limit the number of output groups}

\item{contamination}{as character string}
}
\value{
a tibble on \emph{print}, a tidyproteomics data-object on \emph{save}
}
\description{
\code{summary()} is an analysis function that computes the protein summary
statistics for a given tidyproteomics data object. This is a \emph{passthrough} function
as it returns the original tidyproteomics data-object.
}
