% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_transform_median_subset.R
\name{transform_factor}
\alias{transform_factor}
\title{helper function for normalizing quantitative data from a tidyproteomics data-object}
\usage{
transform_factor(data, data_factor = NULL, ...)
}
\arguments{
\item{data}{tidyproteomics data object}

\item{data_factor}{tidyproteomics data object}

\item{...}{pass through arguments}
}
\value{
a tibble
}
\description{
helper function for normalizing quantitative data from a tidyproteomics data-object
}
