% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_transform_log2.R
\name{transform_log2}
\alias{transform_log2}
\title{helper function for normalizing a quantitative table}
\usage{
transform_log2(table, values = "abundance")
}
\arguments{
\item{table}{a tibble}

\item{values}{a character string}
}
\value{
a tibble
}
\description{
helper function for normalizing a quantitative table
}
