% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_transform_median_subset.R
\name{transform_median}
\alias{transform_median}
\title{helper function for normalizing quantitative data from a tidyproteomics data-object}
\usage{
transform_median(data, group_by = c("identifier"), rename = "log2_med")
}
\arguments{
\item{data}{tidyproteomics data object}

\item{group_by}{character vector}

\item{rename}{character string}
}
\value{
a tibble
}
\description{
helper function for normalizing quantitative data from a tidyproteomics data-object
}
