% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_export.R
\name{write_local}
\alias{write_local}
\title{Helper functio to write data table locally}
\usage{
write_local(table = NULL, file_name = NULL)
}
\arguments{
\item{table}{a tibble}

\item{file_name}{a tibble}
}
\value{
tidyproteomics data object
}
\description{
\code{write_local()} will save the data table in the local project,
}
