% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_subset.R
\name{down_select}
\alias{down_select}
\title{Helper function to subset a data frame}
\usage{
down_select(
  table = NULL,
  variable = NULL,
  operator = c("<", "<=", ">", ">=", "==", "!=", "\%like\%", "\%!like\%"),
  value = NULL
)
}
\arguments{
\item{table}{a tibble}

\item{variable}{a character string}

\item{operator}{a character string}

\item{value}{a character string}
}
\value{
a tibble
}
\description{
Helper function to subset a data frame
}
