% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_enrichment.R
\name{enrichment}
\alias{enrichment}
\title{Compute protein enrichment}
\usage{
enrichment(data = NULL, ..., .term = NULL, .method = c("gsea", "wilcoxon"))
}
\arguments{
\item{data}{tidyproteomics data object}

\item{...}{two sample comparison e.g. experimental/control}

\item{.term}{a character string referencing ".term" in the annotations table}

\item{.method}{a character string}
}
\value{
a tibble
}
\description{
\code{enrichment()} is an analysis function that computes the protein summary
statistics for a given tidyproteomics data object.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)

# using the default GSEA method
hela_proteins \%>\%
   expression(knockdown/control) \%>\%
   enrichment(knockdown/control, .term = "biological_process") \%>\%
   export_analysis(knockdown/control, .analysis = "enrichment", .term = "biological_process")

# using a Wilcoxon Rank Sum method
hela_proteins \%>\%
   expression(knockdown/control) \%>\%
   enrichment(knockdown/control, .term = "biological_process", .method = "wilcoxon") \%>\%
   export_analysis(knockdown/control, .analysis = "enrichment", .term = "biological_process")

}
