% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_export.R
\name{export_analysis}
\alias{export_analysis}
\title{Export the quantitative data from an tidyproteomics data-object}
\usage{
export_analysis(data = NULL, ..., .analysis = NULL, .file_name = NULL)
}
\arguments{
\item{data}{tidyproteomics data object}

\item{file_name}{character string vector}

\item{raw_data}{a boolean}

\item{normalized}{a boolean}

\item{scaled}{a boolean}
}
\value{
a tibble
}
\description{
\code{export()} returns the main quantitative data object as a tibble with
\emph{identifier} as the designation for the measured observation.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)
ecoli_proteins \%>\%
   expression(ko/wt) \%>\%
   export_analysis(ko/wt, .analysis = "expression", .file_name = "ecoli_expression_ko-wt.xlsx")

}
