% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_export.R
\name{export_analysis}
\alias{export_analysis}
\title{Export the quantitative data from an tidyproteomics data-object}
\usage{
export_analysis(
  data = NULL,
  ...,
  .analysis = NULL,
  .term = NULL,
  .append = NULL,
  .file_name = NULL
)
}
\arguments{
\item{data}{tidyproteomics data object}

\item{...}{two sample comparison e.g. experimental/control}

\item{.analysis}{a character string for the specific analysis to export. For
example, the base analysis 'counts' always exists, it is the base analysis
supporting plot_counts(). The other analysis are 'expression' and 'enrichment',
which are only available when those analyses have been performed.}

\item{.term}{a character string of the term from an enrichment analysis. Use
the show_annotations() function to list the available terms.}

\item{.append}{a character string of the term to append to the output. Use
the show_annotations() function to list the available terms.}

\item{.file_name}{a character string for file to write to, format implied
from string ('.rds', '.xlsx', '.csv', '.tsv')}
}
\value{
a tibble
}
\description{
\code{export_analysis()} returns the main quantitative data object as a tibble with
\emph{identifier} as the designation for the measured observation.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)
hela_proteins \%>\%
   expression(knockdown/control) \%>\%
   export_analysis(knockdown/control,
                   .analysis = "expression")

hela_proteins \%>\%
   export_analysis(.analysis = "counts")

}
