% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_import.R
\name{import}
\alias{import}
\title{Main function for importing data to the tidyproteomics list object}
\usage{
import(files = NULL, platform = NULL, analyte = NULL)
}
\arguments{
\item{files}{a character vector of file paths}

\item{platform}{the source of the data (ProteomeDiscoverer, MaxQuant)}

\item{analyte}{the omics analyte (proteins, peptides)}
}
\value{
a tidyproteomics list data-object
}
\description{
Main function for importing data to the tidyproteomics list object
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)
hela_proteins <- path_to_package_data("p97KD_HCT116") \%>\%
   import("ProteomeDiscoverer", "proteins")

hela_proteins \%>\% summary("sample")

}
