% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_imputation.R
\name{impute}
\alias{impute}
\title{Main method for imputing missing values}
\usage{
impute(
  data = NULL,
  impute_function = base::min,
  method = c("within", "between"),
  minimum_to_impute = 0.25,
  cores = 2
)
}
\arguments{
\item{data}{a tidyproteomics list data-object}

\item{impute_function}{summary statistic function. Default is base::min, examples of
other functions include min, max, mean, sum. Note, NAs will be
be removed in the function call.}

\item{method}{a character string to indicate the imputation method (within, between)}

\item{minimum_to_impute}{the minimum ratio to impute at}

\item{cores}{the number of threads used to speed the calculation}
}
\value{
a tidyproteomics list data-object
}
\description{
Main method for imputing missing values
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)
hela_proteins \%>\% summary("sample")

hela_proteins \%>\% impute(impute_function = stats::median) \%>\% summary("sample")

hela_proteins \%>\% impute(impute_function = "randomforest") \%>\% summary("sample")

}
