% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_byfunction.R
\name{impute_byfunction}
\alias{impute_byfunction}
\title{Imputes missing values based on a given function}
\usage{
impute_byfunction(
  table = NULL,
  impute_function = stats::median,
  method = c("within", "between"),
  minimum_to_impute = 0.25
)
}
\arguments{
\item{table}{a tibble}

\item{impute_function}{summary statistic function. Default is median, examples of
other functions include min, max, mean, sum.}

\item{method}{a character string to indicate the imputation method (within, between)}

\item{minimum_to_impute}{the minimum ratio to impute at}
}
\value{
a tibble
}
\description{
Imputes missing values based on a given function
}
