% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_merge.R
\name{merge}
\alias{merge}
\title{Merge multiple tidyproteomics data-objects}
\usage{
merge(data_list = NULL, quantitative_source = c("raw", "selected", "all"))
}
\arguments{
\item{data_list}{a list of tidyproteomics data objects}

\item{quantitative_source}{a character string indicating which quantitative
value to merge on. If \code{selected} is chosen then each dataset's specific normalization
will be used and renamed to 'abundance_selected'. If \code{all} is chosen, then the
possibility exists that some normalization values will fillin with NAs.}
}
\value{
a tidyproteomics data object
}
\description{
\code{merge()} returns a single tidyproteomics data object from multiple.
}
