% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano plot of expression values}
\usage{
plot_volcano(
  data = NULL,
  ...,
  log2fc_min = 2,
  log2fc_column = "log2_foldchange",
  significance_max = 0.05,
  significance_column = "adj_p_value",
  labels_column = "protein",
  show_pannels = TRUE,
  show_lines = TRUE,
  show_fc_scale = TRUE,
  point_size = NULL,
  color_positive = "dodgerblue",
  color_negative = "firebrick1",
  destination = "plot",
  height = 5,
  width = 8
)
}
\arguments{
\item{data}{a tibble}

\item{...}{two sample comparison}

\item{log2fc_min}{a numeric defining the minimum log2 foldchange to highlight.}

\item{log2fc_column}{a character defining the column name of the log2 foldchange values.}

\item{significance_max}{a numeric defining the maximum statistical significance to highlight.}

\item{significance_column}{a character defining the column name of the statistical significance values.}

\item{labels_column}{a character defining the column name of the column for labeling.}

\item{show_pannels}{a boolean for showing colored up/down expression panels.}

\item{show_lines}{a boolean for showing threshold lines.}

\item{show_fc_scale}{a boolean for showing the secondary foldchange scale.}

\item{point_size}{a numeric for changing the point size.}

\item{color_positive}{a character defining the color for positive (up) expression.}

\item{color_negative}{a character defining the color for negative (down) expression.}

\item{destination}{a character string}

\item{height}{a numeric}

\item{width}{a numeric}
}
\value{
a ggplot2 object
}
\description{
\code{plot_volcano()} is a GGplot2 implementation for plotting the expression differences
as foldchange ~ statistical significance. See also \code{plot_proportion()}.  This function can
take either a tidyproteomics data object or a table with the required headers.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)
hela_proteins \%>\%
   expression(knockdown/control) \%>\%
   plot_volcano(knockdown/control, log2fc_min = 0.5, significance_column = "p_value")

# generates the same out come
# hela_proteins \%>\%
#     expression(knockdown/control) \%>\%
#     export_analysis(knockdown/control, .analysis = "expression") \%>\%
#     plot_volcano(log2fc_min = 0.5, significance_column = "p_value")

# display the gene name instead
hela_proteins \%>\%
   expression(knockdown/control) \%>\%
   plot_volcano(knockdown/control, log2fc_min = 0.5, significance_column = "p_value", labels_column = "gene_name")

}
