% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_object.R
\name{print.tidyproteomics}
\alias{print.tidyproteomics}
\title{Tidy-Quant data object print definition}
\usage{
\method{print}{tidyproteomics}(obj)
}
\arguments{
\item{obj}{tidyproteomics data object}
}
\value{
print object summary
}
\description{
Tidy-Quant data object print definition
}
