% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reassign.R
\name{reassign}
\alias{reassign}
\title{reassign the sample info}
\usage{
reassign(data = NULL, ..., .replace = NULL)
}
\arguments{
\item{data}{a tidyproteomics data-object}

\item{...}{a three part expression (eg. x == a)}

\item{.replace}{a character string}
}
\value{
a tidyproteomics data-object
}
\description{
\code{reassign()} enables editing of the sample descriptive in the experimental table.
This function will only replace the sample string and update the replicate number.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)

# check the experiment table
hela_proteins \%>\% summary("experiment")

# make the modification
hela_proteins \%>\%
   reassign(sample == "control", .replace = "ct") \%>\%
   reassign(sample == "knockdown", .replace = "kd") \%>\%
   summary("sample")

# reassign specific file_ids
hela_proteins \%>\%
   reassign(sample_file == "f1", .replace = "new") \%>\%
   reassign(sample_file == "f2", .replace = "new") \%>\%
   summary("sample")

}
