% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_reverselog_transform.R
\name{reverselog_transformation}
\alias{reverselog_transformation}
\title{Reverse the plot axis for log transformation}
\usage{
reverselog_transformation(base = exp(1))
}
\arguments{
\item{base}{a numeric}
}
\value{
a ggplot scale transformation
}
\description{
Reverse the plot axis for log transformation
}
