% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_stats.R
\name{summary.tidyproteomics}
\alias{summary.tidyproteomics}
\title{Summarize the data}
\usage{
\method{summary}{tidyproteomics}(
  data,
  by = c("global"),
  destination = c("print", "save", "return"),
  limit = 25,
  contamination = NULL
)
}
\arguments{
\item{data}{tidyproteomics data object}

\item{by}{what to summarize}

\item{destination}{character string, one of (save, print)}

\item{contamination}{as character string}
}
\value{
a tibble on \emph{print}, a tidyproteomics data-object on \emph{save}
}
\description{
\code{summary()} is an analysis function that computes the protein summary
statistics for a given tidyproteomics data object. This is a \emph{passthrough} function
as it returns the original tidyproteomics data-object.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(tidyproteomics)

# a global summary
hela_proteins \%>\% summary()

# a summary by sample
hela_proteins \%>\% summary("sample")

# a summary by sample with match_between_runs removed
hela_proteins \%>\%
   subset(match_between_runs == FALSE) \%>\%
   summary("sample")

# a summary of match_between_runs
hela_proteins \%>\% summary("match_between_runs")

hela_proteins \%>\% summary("cellular_component")

hela_proteins \%>\% summary("biological_process")

}
