% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_subset.R
\name{tidyproteomics_quo}
\alias{tidyproteomics_quo}
\title{Helper function to subset a data frame}
\usage{
tidyproteomics_quo(...)
}
\arguments{
\item{table}{a tibble}

\item{variable}{a character string}

\item{value}{a character string}

\item{operator}{a character string}
}
\value{
a tibble
}
\description{
Helper function to subset a data frame
}
